Updated Relationships, Created Seed method, seed helper, Populated database

+++ Pantry/DataModels/AccountIngredient.cs
@@ -18,8 +18,6 @@ namespace Pantry.Data.Models
         public virtual Account Account { get; set; }
         public virtual Ingredient Ingredient { get; set; }
 
-
-        //base unit is milliliters
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
     }
 }
